requirements.txt
streamlit==1.31.0
pandas==2.1.0
openpyxl==3.1.2
pillow==10.0.0
matplotlib==3.7.3
streamlit-authenticator==0.2.3
plotly==5.18.0
pyyaml==6.0.1 